+++ Makefile
@@ -44,8 +44,7 @@ help:
 	@echo "make test-snapshots                  - runs test, updating snapshots"
 	@echo "make clean                           - removes node_modules and built artifacts"
 	@echo "----------------------- CI Commands  -------------------------"
-	@echo "make build                           - builds a bundle with production settings"
-	@echo "make build_and_deploy                - builds and deploys the production bundle"
+	@echo "make build-production                - builds a bundle with production settings"
 
 
 # ---- YVM Command Stuff ----
@@ -59,19 +58,15 @@ install-watch: node_modules
 	scripts/install-watch.sh
 
 
-# ---- Infrastructure for Test/Deploy ----
+# ---- Webpack ----
 
-.PHONY: build
+.PHONY: build-production
 build: node_modules
-	$(WEBPACK) --config webpack/webpack.config.base.js
+	$(WEBPACK) --config webpack/webpack.config.production.js
 
 .PHONY: build-dev
 build-dev: node_modules
-	$(WEBPACK) --config webpack/webpack.config.dev.js
-
-.PHONY: build_and_deploy
-build_and_deploy: node_modules
-	$(WEBPACK) --config webpack/webpack.config.deploy.js
+	$(WEBPACK) --config webpack/webpack.config.development.js
 
 
 # -------------- Linting --------------
