+++ package.json
@@ -26,7 +26,7 @@
     "commander": "^2.15.1",
     "cosmiconfig": "^5.0.5",
     "fs-extra": "^7.0.1",
-    "gh-release": "^3.2.1",
+    "gh-release": "^3.4.0",
     "request": "^2.87.0",
     "shelljs": "^0.8.2",
     "targz": "^1.0.1"
