+++ webpack/webpack.config.production.js
@@ -1,0 +1,11 @@
+const ZipPlugin = require('zip-webpack-plugin')
+
+const baseConfig = require('./webpack.config.base')
+
+baseConfig.plugins.push(
+    new ZipPlugin({
+        filename: 'yvm.zip',
+    }),
+)
+
+module.exports = baseConfig
